Moved more private files out of the root directory
