+++ list.php
@@ -1,5 +1,5 @@
 <?php
-require "header.php";
+require "scripts/pi-hole/php/header.php";
 
 $list = $_GET['l'];
 
@@ -52,7 +52,7 @@ function getFullName() {
 <ul class="list-group" id="list"></ul>
 
 <?php
-require "footer.php";
+require "scripts/pi-hole/php/footer.php";
 ?>
 
 <script src="scripts/pi-hole/js/list.js"></script>
