+++ scripts/pi-hole/php/footer.php
